requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
attrs==23.2.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.3.2
certifi==2025.10.5
cffi==1.16.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.4
duckdb==1.4.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.2.4
httplib2==0.22.0
idna==3.11
isodate==0.6.1
itsdangerous==2.2.0
javascript==1!1.1.2
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nextcord==2.6.0
oauthlib==3.2.2
packaging==23.2
propcache==0.4.1
protobuf==4.25.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.1.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==1.3.1
rsa==4.9
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing-inspect==0.9.0
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
